libipl: Run HWP only if the target is enabled
Use istep pre function to probe the required targets.  Only probed
targets get enabled, so running HWP on enabled targets matches the
intent of running HWP on required targets.

This ensures that the logic to pick the required targets is in a
single place and not distributed all over the HWP code.

Signed-off-by: Amitay Isaacs <amitay@ozlabs.org>