genesisboot: Create parent directory if it does not exist
The c++ "create_directories" api will return true only when it able
to create the given path else it will return false. so, it failing
in following case i.e when parent directory exists but, genesisboot
file is not present, then it will trigger to update all the requried
targets HWAS state and after that it will try to create genesisboot
file but it failing in create_directories call due to the parent
directory exists and it not able to create parent directory.

Signed-off-by: Ramesh Iyyar <rameshi1@in.ibm.com>