libipl: Checking write flag before applying guard
Change:
Adding checkWriteFlag, which will read the write flag
in guard partition, this to see if the guard partition
is not corrupt or not. If the write flag is set to 0x00
checkWriteFlag will return true and need to clear the
guard records from the partition.

Tested:
Setting the write flag as 0x00, seeing that guard records
are getting clear.

Signed-off-by: Chirag Sharma <chirshar@in.ibm.com>