libipl: ATTR_SYS_CLOCK_DECONFIG_STATE based clock state update
ATTR_SYS_CLOCK_DECONFIG_STATE contains the functional state
of the system clocks/oscillators. This is the way that Hostboot
communicates to the BMC which clocks have failed.
Hostboot does not model the clocks as targets, which is why
we need a new attribute to manage clock state.

This attribute is set by Hostboot and read by the BMC on
reconfig loops so that it can tell which clocks are functional.
See the enumeration SYS_CLOCK_DECONFIG_STATE for the possible
values of this attribute.

A_DECONFIG means OSC1 is active, B_DECONFIG means OSC0 is active.
"A" corresponds to the CP_REFCLOCK_SELECT value of OSC0 and
"B" to OSC1

Signed-off-by: Jayanth Othayoth <ojayanth@in.ibm.com>