libphal : add new error type SBE is in the invalid state
Adding new error type SBE_INVALID_STATE, this is used in places
where SBE is not in the expected state.

During SBE error handling we expect the system to be either in
DEBUG/FAILED state, if SBE is not in the expected state
SBE_INVALID_STATE error will be returned.

Signed-off-by: Marri  Rao <devenrao@in.ibm.com>
Change-Id: I0d7e0bc003e0a571cabf3f62b86f6b3458ffab34