libphal : initial commit for sbe return code handling
For the primary processor, there will be primary and backup
boot and measurement seeprom hosting SBE code.

For any sbe chip-op failure SBE FFDC will be collected to
determine the cause.

If SBE is in a bad state(dead /not communicating) SBE FFDC
chip-op fails with a TIMEOUT error.

If SBE is in a bad state execute p10_extract_sbe_rc HWP, which
uses CFAM registers to determine the SBE failure cause and
recovery action.

Recovery action could be reattempt by re-ipling from current
seeprom side, switch to backup seeprom if fail to recover
from primary seeprom.

This initial commit caters to writing stubs for the different
recovery actions supported for the recovery of failed SBE.

Signed-off-by: Marri  Rao <devenrao@in.ibm.com>
Change-Id: Ibc01ec904bbdff64e283b3de754db43a12360148