libipl: Add support to log message into journal
If there is no external log callback setup, the logs were printed on the
terminal. But there was no way to look at them later to debug

Use case:
When user executes ecmd istep command line, ipl traces were not logged to the
journal and it is difficult to trace any error. ecmd-pdbg does not have support
to register callbacks in ipl library as it is command line tool

Now with the compile time flag --enable-journal the logs will be added to
journal with the LOG_INFO priority. If the compile time flag is NOT provided,
then the logs are printed onto the terminal.

Signed-off-by: Deepa Karthikeyan <deepakala.karthikeyan@ibm.com>