libphal: Add method to find if control transitioned to host
Utilizing the xyz.openbmc_project.State.Host interface to ascertain the
host's operational status becomes less accurate when transitioning from
hostboot to host. During the period when host is initializing and not
yet fully operational, the host state is erroneously presumed to be in
hostboot. In situations where host encounters initial boot difficulties,
and the watchdog time is hit as the boot up process has not completed
within a specific time, this watchdog misinterprets the issue as
a problem with hostboot.

To address this, there exists a core scratch register that undergoes an
update by hostboot just before transferring control to host. We have
devised a method that leverages this register to determine whether the
transition to host has already occurred.

By implementing this functionality we can determine which booting
subsystem is failed or stopped responding, and the dump can be extracted
from the right subsystem.

Signed-off-by: Deepa Karthikeyan <deepakala.karthikeyan@ibm.com>