Merging missing commits in open-power/ipl
In OpenBMC 1050 branch for some unavoidable reasons we have
to create a mirror of open-power/ipl repo under ibm-openbmc.
But that was a temporary solution and ultimately we need to
move back to open-power/ipl in 1060 branch. This commit merges
the missing/exclusive commits from ibm-openbmc/ipl repo into
the open-power/ipl repo so that we can move back to original
open-power/ipl repo in OpenBMC 1060 branch.

Missign commits which are merged are:
8c2272cc5a9030acc459e0ccb3f5e561511d344b from ibm-openbmc/ipl-1050
033bf30ce7a906e08d96790434174f706f607fb4 from ibm-openbmc/ipl-1050
(This one was ported to 1020 and 1030 as well)

The commits which were already availbale in open-power/ipl-main are:
ec7d067c7dec06f61d92d186a23e9e8fe1ee4c99 from ibm-openbmc/ipl-1020
6a4a7696a94c06e34fece077236861d291df8508 from ibm-openbmc/ipl-1020
9e84ef27c631457f829166b383c64c9d90c792d4 from ibm-openbmc/ipl-1020
(Actually it combines the above two)

4882ca795c0e69b606860b79cc91d03f834ec92e from ibm-openbmc/ipl-1020

Signed-off-by: Swarnendu Roy Chowdhury <swarnendu.roy.chowdhury@ibm.com>