OdySsey SBE dump support-Commit-1
A table format along with necessary structures added.

We have now OdySsey chip along with the existing p10 which
supports SBE dump collection. In future we may have more
varities of chips. So to make life easy a table format (map)
has been introduced along with necessary structures so that
in future also if a new type of chips is introduced we can
handle that as well with minimum code changes.

Signed-off-by: Swarnendu Roy Chowdhury <swarnendu.roy.chowdhury@ibm.com>