libphal: PHAL API for returning FRU Type
We need to capture the DRAM manufacturer ID in the PEL (error log)
whenever a DIMM is callout/guarded/deconfigured. This is essential
for identifying the DRAM manufacturer since IBM sources DRAM from
different vendors.

The VPD will store the DRAM manufacturer ID in the DIMM inventory
using the VINI DBus interface. The PEL daemon must retrieve this
VPD keyword and add it to a new user data section.

The PEL daemon is unable to determine if a
callout/guarded/deconfigured location code corresponds to a DIMM.
Therefore, PHAL needs to offer an API that can return the FRU type
by utilizing the PHAL device tree.

Change-Id: If31bbe18a36aeb3e5d8844643a5bf00db1515106
Signed-off-by: SwethaParasa <parasa.swetha1@ibm.com>